Rename package ...tshirtcannon.cannons to tshirtcannon.commands and refactor.
Simon made a mistake and made the package for the cannon commands to be cannons,
just a simple mistake and this is a simple fix.
